requirements.txt
# Speech Recognition & Processing
faster-whisper
pyannote.audio
torch
torchaudio

# UI & Config
streamlit
pyyaml
python-dotenv

# LLM
openai
google-generativeai

# Utilities
numpy